Better check_pid implementation from psutil

+++ script/testing/util/test_server.py
@@ -6,6 +6,7 @@ import subprocess
 import sys
 import time
 import traceback
+import errno
 from util import constants
 from util.common import run_command
 
@@ -101,7 +102,7 @@ class TestServer:
             try:
                 s.connect((self.db_host, int(self.db_port)))
                 s.close()
-                print("Connected to server in {} seconds".format(i * 0.1))
+                print("Connected to server in {} seconds [PID={}]".format(i * 0.1, self.db_process.pid))
                 is_db_running = True
                 break
             except:
@@ -173,11 +174,28 @@ class TestServer:
             self.print_output(self.db_output_file)
         return ret_val
     
-    def check_pid(self, pid):        
-        """ Check For the existence of a unix pid. """
+    def check_pid(self, pid):
+        """Check whether pid exists in the current process table."""
+        
+        # Copied from psutil
+        # https://github.com/giampaolo/psutil/blob/5ba055a8e514698058589d3b615d408767a6e330/psutil/_psposix.py#L28-L53
+        
+        if pid == 0:
+            return True
         try:
             os.kill(pid, 0)
-        except OSError:
-            return False
+        except OSError as err:
+            if err.errno == errno.ESRCH:
+                # ESRCH == No such process
+                return False
+            elif err.errno == errno.EPERM:
+                # EPERM clearly means there's a process to deny access to
+                return True
+            else:
+                # According to "man 2 kill" possible error values are
+                # (EINVAL, EPERM, ESRCH) therefore we should never get
+                # here. If we do let's be explicit in considering this
+                # an error.
+                raise err
         else:
             return True
